requirements.txt
elasticsearch==2.4.0
elasticsearch-dsl==2.1.0
dateutils==0.6.6
